+++ doc/changelog.rst
@@ -4,10 +4,38 @@ Changelog
 Like any other piece of software, JsHamcrest is evolving at each release.
 Here you can track our progress:
 
-**Version 0.6.4** (May 25, 2011)*
+**Version 0.7.0** *(Jan 26, 2013)*
+
+* Added matchers :meth:`JsHamcrest.Matchers.equivalentMap` and :meth:`JsHamcrest.Matchers.equivalentArray`;
+
+
+**Version 0.6.8** *(Nov 30, 2012)*
+
+* Added Node.js support;
+
+
+**Version 0.6.7** *(Sep 11, 2011)*
+
+* Updated :meth:`JsHamcrest.Matchers.hasSize` function to make it work with
+  objects;
+
+
+**Version 0.6.6** *(Jun 23, 2011)*
+
+* Fixed broken logic of :meth:`JsHamcrest.areArraysEqual` function;
+
+
+**Version 0.6.5** *(Jun 20, 2011)*
+
+* Description of the Function literal now includes function name when
+  available;
+
+
+**Version 0.6.4** *(May 25, 2011)*
 
 * Updated ``hasMember`` matcher to make it possible to also match considering
   the member's value;
+
 
 **Version 0.6.3** *(Apr 23, 2011)*